# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Patient" }
    post:
      operationId: createPatient
      tags: [patients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePatientBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
    patch:
      operationId: updatePatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePatientBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
  /patients/{id}/timeline:
    get:
      operationId: getPatientTimeline
      summary: Get patient medical timeline
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A chronological list of patient events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelineEvent'
...
  /medications:
    get:
      operationId: listMedications
      tags: [pharmacy]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Medication" }
    post:
      operationId: createMedication
      tags: [pharmacy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMedicationBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Medication" }
  /medications/{id}:
    patch:
      operationId: updateMedication
      tags: [pharmacy]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMedicationBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Medication" }
  /pharmacy-sales:
    get:
      operationId: listPharmacySales
      tags: [pharmacy]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PharmacySale" }
    post:
      operationId: createPharmacySale
      tags: [pharmacy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePharmacySaleBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PharmacySale" }
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Staff" }
    post:
      operationId: createStaff
      tags: [staff]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateStaffBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStaffBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
...
  /radiology-orders:
    get:
      operationId: listRadiologyOrders
      tags: [radiology]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RadiologyOrder" }
    post:
      operationId: createRadiologyOrder
      tags: [radiology]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRadiologyOrderBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RadiologyOrder" }
  /radiology-orders/{id}:
    patch:
      operationId: updateRadiologyOrder
      tags: [radiology]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateRadiologyOrderBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RadiologyOrder" }
...
components:
  schemas:
...
    Staff:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        username: { type: string }
        role: { type: string, description: "admin|doctor|nurse|pharmacist|cashier|labtech|radtech|manager" }
        department: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        salary: { type: ["number", "null"] }
        joinedAt: { type: string, format: date-time }
      required: [id, fullName, username, role, joinedAt]
    CreateStaffBody:
      type: object
      properties:
        fullName: { type: string }
        username: { type: string }
        password: { type: string }
        role: { type: string }
        department: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        salary: { type: ["number", "null"] }
      required: [fullName, username, password, role]
    UpdateStaffBody:
      type: object
      properties:
        fullName: { type: string }
        department: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        salary: { type: ["number", "null"] }
        role: { type: string }
    Patient:
      type: object
      properties:
        id: { type: integer }
        mrn: { type: string, description: "Medical Record Number" }
        fullName: { type: string }
        gender: { type: string, description: "male|female" }
        dob: { type: ["string", "null"], format: date }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        bloodType: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, mrn, fullName, gender, createdAt]
    CreatePatientBody:
      type: object
      properties:
        fullName: { type: string }
        gender: { type: string }
        dob: { type: ["string", "null"], format: date }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        bloodType: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [fullName, gender]
    UpdatePatientBody:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        bloodType: { type: ["string", "null"] }
        emergencyContact: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...
    RadiologyOrder:
      type: object
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: string }
        doctorId: { type: integer }
        doctorName: { type: string }
        modality: { type: string, description: "xray|ct|mri|ultrasound" }
        bodyPart: { type: string }
        requestedAt: { type: string, format: date-time }
        status: { type: string, description: "pending|completed" }
        report: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        price: { type: number }
      required: [id, patientId, patientName, doctorId, doctorName, modality, bodyPart, requestedAt, status, price]
    CreateRadiologyOrderBody:
      type: object
      properties:
        patientId: { type: integer }
        doctorId: { type: integer }
        modality: { type: string }
        bodyPart: { type: string }
        price: { type: number }
      required: [patientId, doctorId, modality, bodyPart, price]
    UpdateRadiologyOrderBody:
      type: object
      properties:
        status: { type: string }
        report: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
    Medication:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        unit: { type: string, description: "tablet|bottle|tube|vial|sachet" }
        stock: { type: integer }
        reorderLevel: { type: integer }
        price: { type: number }
        expiresOn: { type: ["string", "null"], format: date }
        manufacturer: { type: ["string", "null"] }
      required: [id, name, unit, stock, reorderLevel, price]
    CreateMedicationBody:
      type: object
      properties:
        name: { type: string }
        category: { type: ["string", "null"] }
        unit: { type: string }
        stock: { type: integer }
        reorderLevel: { type: integer }
        price: { type: number }
        expiresOn: { type: ["string", "null"], format: date }
        manufacturer: { type: ["string", "null"] }
      required: [name, unit, stock, reorderLevel, price]
    UpdateMedicationBody:
      type: object
      properties:
        stock: { type: integer }
        reorderLevel: { type: integer }
        price: { type: number }
        expiresOn: { type: ["string", "null"], format: date }
        manufacturer: { type: ["string", "null"] }
    PharmacySale:
      type: object
      properties:
        id: { type: integer }
        patientName: { type: string }
        patientId: { type: ["integer", "null"] }
        medicationId: { type: integer }
        medicationName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        total: { type: number }
        soldAt: { type: string, format: date-time }
      required: [id, patientName, medicationId, medicationName, quantity, unitPrice, total, soldAt]
    CreatePharmacySaleBody:
      type: object
      properties:
        patientName: { type: string }
        patientId: { type: ["integer", "null"] }
        medicationId: { type: integer }
        quantity: { type: integer }
      required: [patientName, medicationId, quantity]
...
    TimelineEvent:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [opd, ipd, lab, radiology, prescription, billing] }
        date: { type: string, format: date-time }
        title: { type: string }
        detail: { type: string }
        status: { type: string }
      required: [id, type, date, title, detail, status]
...